* refactor sensor simulation modules to one module called SensorSimManager. allows better failure handling * still all sensors can be enabled/disabled through parameters * check airspeed simulation (sih.cpp has seperate calc.)
