+++ ROMFS/px4fmu_common/init.d/rcS
@@ -336,10 +336,7 @@ else
 		if param compare SYS_HITL 2
 		then
 			simulator_sih start
-			sensor_baro_sim start
-			sensor_mag_sim start
-			sensor_gps_sim start
-			sensor_agp_sim start
+			sensor_sim_manager start
 		fi
 
 	else
