+++ boards/px4/sitl/default.px4board
@@ -55,7 +55,7 @@ CONFIG_COMMON_SIMULATION=y
 CONFIG_MODULES_SIMULATION_GZ_MSGS=y
 CONFIG_MODULES_SIMULATION_GZ_BRIDGE=y
 CONFIG_MODULES_SIMULATION_GZ_PLUGINS=y
-CONFIG_MODULES_SIMULATION_SENSOR_AGP_SIM=y
+CONFIG_MODULES_SIMULATION_SENSOR_SIM_MANAGER=y
 CONFIG_MODULES_TEMPERATURE_COMPENSATION=y
 CONFIG_MODULES_UUV_ATT_CONTROL=y
 CONFIG_MODULES_UUV_POS_CONTROL=y
