+++ src/modules/simulation/Kconfig
@@ -5,10 +5,7 @@ menu "Simulation"
         depends on PLATFORM_POSIX
         select MODULES_SIMULATION_BATTERY_SIMULATOR
         select MODULES_SIMULATION_PWM_OUT_SIM
-        select MODULES_SIMULATION_SENSOR_AIRSPEED_SIM
-        select MODULES_SIMULATION_SENSOR_BARO_SIM
-        select MODULES_SIMULATION_SENSOR_GPS_SIM
-        select MODULES_SIMULATION_SENSOR_MAG_SIM
+        select MODULES_SIMULATION_SENSOR_SIM_MANAGER
         select MODULES_SIMULATION_SYSTEM_POWER_SIMULATOR
         select MODULES_SIMULATION_SIMULATOR_MAVLINK
         select MODULES_SIMULATION_SIMULATOR_SIH
