--- src/modules/simulation/sensor_agp_sim/Kconfig
@@ -1,5 +1,0 @@
-menuconfig MODULES_SIMULATION_SENSOR_AGP_SIM
-	bool "sensor_agp_sim"
-	default n
-	---help---
-		Enable support for sensor_agp_sim
