--- src/modules/simulation/sensor_airspeed_sim/Kconfig
@@ -1,5 +1,0 @@
-menuconfig MODULES_SIMULATION_SENSOR_AIRSPEED_SIM
-	bool "sensor_airspeed_sim"
-	default n
-	---help---
-		Enable support for sensor_airspeed_sim
