--- src/modules/simulation/sensor_gps_sim/Kconfig
@@ -1,5 +1,0 @@
-menuconfig MODULES_SIMULATION_SENSOR_GPS_SIM
-	bool "sensor_gps_sim"
-	default n
-	---help---
-		Enable support for sensor_gps_sim
