+++ src/modules/simulation/sensor_sim_manager/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2021 PX4 Development Team. All rights reserved.
+# Copyright (c) 2025 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -32,14 +32,17 @@
 ############################################################################
 
 px4_add_module(
-	MODULE modules__simulation__senosr_mag_sim
-	MAIN sensor_mag_sim
+	MODULE modules__simulation__sensor_sim_manager
+	MAIN sensor_sim_manager
 	COMPILE_FLAGS
+	INCLUDES
+		${CMAKE_CURRENT_SOURCE_DIR}/../failure_injection
 	SRCS
-		SensorMagSim.cpp
-		SensorMagSim.hpp
+		SensorSimManager.cpp
+		SensorSimManager.hpp
+		${CMAKE_CURRENT_SOURCE_DIR}/../failure_injection/failureInjection.cpp
 	DEPENDS
-		drivers_magnetometer
 		geo
+		mathlib
 		px4_work_queue
-	)
+)
